nfe-monitor_app/requirements.txt
Flask
Flask-SQLAlchemy
lxml
pandas
python-dotenv